computer-use-demo/computer_use_demo/requirements.txt
streamlit>=1.38.0
anthropic[bedrock,vertex]>=0.37.1
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
pyautogui>=0.9.54
mss>=9.0.1
pillow>=10.2.0